deepspeed-chat: add end-of-text special token
Stages 1 & 2 append '<|endoftext|>' text marker to all samples.
However, some tokenizers (e.g. OPT, Bloom), encode this marker as a sequence
of subword tokens and not as a single special token.

This commit adds an optional support to add the EOT marker as a special token
to force the tokenizer to encode it as a single token.

Note that using EOT special token may change the dynamics of stage3 training.
Therefore, to be backward compliant, this commit makes it optional.

Change-Id: If98d348fcaa7d6685e755aabe305e23e7649c367
Signed-off-by: Moshe Island <misland@habana.ai>