Filter transformers version 4.33.2 due to bug (#735)
Performance metrics in DeepSpeed-Chat were broken with transformers version 4.33.2, so that version is explicitly filtered out in the DeepSpeed-Chat requirements.txt. This will fix the currently broken nv-ds-chat workflow.

The issue is recently fixed in this HuggingFace transformers PR, so this will not be a problem in the next transformers release.